requirements.txt
django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-phone-field==1.8.1
black==22.6.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
python-dotenv==0.20.0